# Weighted learning of TIs and HNs