**What this pull request is trying to achieve.**
---

This will add the ability to train Embeddings and Hypernetworks using a weighted loss.
It will add the option to load PNG alpha channels as weight masks that will be used during loss calculation, in effect focusing the attention towards the masked subject.

During my training I've found that focusing the attention in this manner results in much faster learning of subjects, as well as learning more detailed features of the subjects.

Example of what is being emphasized during learning. (_Regular_ is for demonstration only)
![normal](https://user-images.githubusercontent.com/1111394/212354978-903533ab-f87d-4715-87d9-c4ab89d03c34.png) ![depth](https://user-images.githubusercontent.com/1111394/212354982-f446dbeb-fb24-4d29-a37b-9c2ef2167d5b.png) ![clipseg](https://user-images.githubusercontent.com/1111394/212354985-d40b7112-da58-494c-b16d-642220af577a.png)
1. Regular - 2. Depth - 3. CLIPSeg

Note that this does not interfere with Dreambooth or regular image generation.
Only during the loss calculation for TIs and HNs is this adjusted loss calculated.

In the future this could be extended to load side-by-side greyscale files, for example depth maps.
I.e. if you have image.png and image.weight.png in the same folder, it will load one as the weight for the other.


**Additional notes and description of changes**
---
I've added an extra hijack that adds a weighted_forward function, next to the regular forward function.
This function will do a regular forward pass, but before doing so, substitute the get_loss function and add the weight tensors to the model so the weighted_loss function can apply it.
Before calculating the mean, it will multiply the individual latent losses with the weight mask, resulting in more or less loss for pixels based on the amount of weight. Afterwards it will return the mean to the training function.

**Environment this was tested in**

 - OS: Windows
 - Browser: Firefox
 - Graphics card: NVIDIA RTX 2080 8GB

**How to use this**
---
1. Preprocess images normally.
2. Choose one of the following:
    a. Generate a depth map using MiDaS
    b. Generate an attention map using CLIPSeg
    c. Manually draw a map
    d. Create a generic gradient with falloff to the image edges.
3. Add the map from step 2 as an alpha channel to the image from step 1.
    a. Using a fairly easy Python script and PILLOW
    b. Manually in Photoshop or GIMP
Note: be sure that the alpha channel is 8-bit, not single color transparency. 

In the future I would like to release a script that would do these steps automatically.
(Perhaps even add MiDaS generation to the preprocess step)

**Screenshots**
---
<img width="264" alt="weighted-learning-ui" src="https://user-images.githubusercontent.com/1111394/212950349-157dbd43-dd71-4ade-aff8-59aafd6ba9d3.png">


**Results**
---
Haven't found time for extensive training runs and comparison images.
But I have received positive feedback from multiple people on Discord.

**Timing:**
No significant impact.

Without:
Loading dataset: 4.70it/s
Training: 2.95it/s

With:
Loading dataset: 4.70it/s
Training: 2.94it/s