@Shondoit 2023-01-14
Some interesting links to get weight maps:
https://github.com/isl-org/MiDaS
https://github.com/danielgatis/rembg
https://github.com/graemeniedermayer/clothseg
https://github.com/xuebinqin/U-2-Net
https://github.com/pymatting/pymatting