@mykeehu 2023-01-14
This is similar to the Attention Map option of the [DreamArtist extension](https://github.com/7eu7d7/DreamArtist-sd-webui-extension), it is only for embeddings, not for HN. Did you look at that?