@Shondoit 2023-01-15
> This is similar to the Attention Map option of the [DreamArtist extension](https://github.com/7eu7d7/DreamArtist-sd-webui-extension), it is only for embeddings, not for HN. Did you look at that?

I have not, but I will check it out. Thanks for the reference.