@FurkanGozukara 2023-01-16
With the default way what learn rate do you suggest to teach a face? @Shondoit