@Shondoit 2023-01-16
> With the default way what learn rate do you suggest to teach a face? @Shondoit
Currently there's no consensus. I've found the default (0.005) to be good enough.