@FurkanGozukara 2023-01-16
@mykeehu in that video how to determine number of vectors is still not explained. he gives some numbers but no backing up how he did come up with those numbers. i really need to understand how vectors are utilized.

also he used image descriptions but when training those prompts are not even used lol :)