@FurkanGozukara 2023-01-16
> In TI training, we use an average of 5 images per vector. You can set up more vectors, in which case the embedding will be stronger. We have talked a lot about the details in [this](https://github.com/AUTOMATIC1111/stable-diffusion-webui/discussions/1528) and [this](https://github.com/AUTOMATIC1111/stable-diffusion-webui/discussions/4914) thread. Keywords are extracted from the original thing, so you have something to compare what the system still sees in the image to. So you can type the keywords in the prompt afterwards, but what's not there is remembered by the training.

it shows last used prompt. in my case it didnt read the image captions generated by pre processing. it just used image name and the model name can you verify that?