@FurkanGozukara 2023-01-16
> @mykeehu
> 
> > In TI training, we use an average of 5 images per vector. You can set up more vectors, in which case the embedding will be stronger.
> 
> This rule of thumb doesn't make sense at all. If you have 10 picture or 1000 pictures, if they're all of the same subject, we want it to learn the same result. It makes more sense to say 1 vector per subject complexity. I.e., a complex subject might need an extra vector, or an extra subject might need 2x the vectors. Keep in mind that "Emma Watson" consists of just 2 vectors and is trained on multiple thousands of images.
> 
> More training steps per image in the dataset does make sense though, cause it needs more time to come to a consensus.

ok now makes more sense. so for a single person we can say 2 is at max is sufficient