@kidol 2023-01-21
**One issue I see with this PR:** The latent sample shape is always the same (I think?), so why not hardcode it and leave the sampling method logic unchanged?

Other than that, I tested this and it works as described. Not used TI or Hypernetworks much recently, so can't say anything about training speed, but it should be obvious that a weighted loss trains faster.

I've used this lib to generate the masks: https://github.com/1adrianb/face-alignment
It can detect faces and extract the landsmarks. landmarks -> binary face hull -> hull expansion -> blur = 4th channel

I guess if this gets merged, people can create extensions for the different mask generation methods, so no need to include a specific method in this PR.