@Shondoit 2023-01-21
> **One issue I see with this PR:** The latent sample shape is always the same (I think?), so why not hardcode it and leave the sampling method logic unchanged?
It's definitely different between 512 and 768 models (64 vs 96).
Also, I'm not entirely sure what happens if you use non-square aspect ratios. But it's good to be future prove; if there ever comes a 1024 model, it won't break.

> I guess if this gets merged, people can create extensions for the different mask generation methods, so no need to include a specific method in this PR.
Yeah, that was the idea from the start. Keep this simple and let people do whatever they want.  Tooling will either be external or extensions. (Or perhaps one very basic one in the preprocessing tab.)