@Shondoit 2023-01-23
@AUTOMATIC1111 Is there something I should do or consider before you want to implement this?

Pre-empting the question "can't this be an extension?", it's 99% based on normal TI/HN learning code. Only the Loss function is tweaked.  I think it would be a waste to copy/paste everything in an extension. And I'm not sure it's easy to hijack only one line of WebUI code in the TI/HN training functions.  Besides, it has no impact for users that don't use the function.