@Nacurutu 2023-01-23
> This is just another feature that can be used for training. It's not magically going to fix everything. That being said, I've found that setting the focus on the face, it trains less on the background, so you need less tricks to coerce the training to train on the face. (I.e., proper captioning, masking out the background, etc.) With this you can mask out the object/person you want to train on and it will focus it's training on that.

Thank you very much for your answer, I'm going to give it a try. Looks promising.