@FurkanGozukara 2023-01-23
> @Shondoit thanks for this... May I ask, with this we get a real improvement when training faces? Im trying to train faces but I'm not getting the results I expect.

you can watch my video i got pretty good results already

[How To Do Stable Diffusion Textual Inversion (TI) / Text Embeddings By Automatic1111 Web UI Tutorial](https://www.youtube.com/watch?v=dNOpWt-epdQ)